# p2p-server/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_message.AckMessagesRequest:
    properties:
      message_ids:
        items:
          type: string
        type: array
    required:
    - message_ids
    type: object
  internal_message.InitialMessageDTO:
    properties:
      sender_ephemeral_key:
        type: string
      sender_identity_key:
        type: string
    type: object
  internal_message.PendingMessageDTO:
    properties:
      ciphertext:
        type: string
      id:
        type: string
      message_type:
        type: integer
      sender_id:
        type: string
      timestamp:
        type: integer
    type: object
  internal_message.PendingMessagesResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/internal_message.PendingMessageDTO'
        type: array
    type: object
  internal_message.StoreMessageRequest:
    properties:
      ciphertext:
        type: string
      initial_message:
        $ref: '#/definitions/internal_message.InitialMessageDTO'
      message_id:
        type: string
      message_type:
        type: integer
      ratchet_index:
        type: integer
      recipient_id:
        type: string
      sender_id:
        type: string
      timestamp:
        type: integer
    required:
    - ciphertext
    - message_id
    - recipient_id
    - sender_id
    - timestamp
    type: object
  internal_message.StoreMessageResponse:
    properties:
      message_id:
        type: string
      status:
        type: string
      will_notify:
        type: boolean
    type: object
  internal_user.SearchUserResponse:
    properties:
      found:
        type: boolean
      user:
        $ref: '#/definitions/internal_user.UserDTO'
    type: object
  internal_user.UserDTO:
    properties:
      id:
        type: string
      identity_public_key:
        type: string
      last_seen:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    name: Triddov
    url: https://github.com/Triddov
  description: |-
    REST API peer-to-peer мессенджера: аутентификация по email-коду,
    управление профилем, обмен оффлайн-сообщениями (E2EE) и распространение
    Signal Protocol prekey-бандлов. Сервер не может расшифровать сообщения.
    WebRTC-сигналинг вынесен в отдельный сервис и здесь не документируется.
  license:
    name: See repository LICENCE
  title: P2P Messenger REST API
  version: "1.0"
paths:
...
  /messages/ack:
    post:
      consumes:
      - application/json
      description: Помечает перечисленные сообщения как доставленные, чтобы они больше
        не выдавались.
      parameters:
      - description: Список ID сообщений
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_message.AckMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Подтвердить доставку сообщений
      tags:
      - Messages
  /messages/pending:
    get:
      description: |-
        Возвращает до 100 ожидающих сообщений текущего пользователя. ciphertext — в base64.
        После сохранения на клиенте их следует подтвердить через /messages/ack.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_message.PendingMessagesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить недоставленные сообщения
      tags:
      - Messages
  /messages/store:
    post:
      consumes:
      - application/json
      description: |-
        Кладёт E2EE-зашифрованное сообщение в очередь получателя. Сервер не может его расшифровать.
        ciphertext передаётся в base64. Используется, когда получатель недоступен по P2P.
      parameters:
      - description: Зашифрованное сообщение
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_message.StoreMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_message.StoreMessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Сохранить сообщение для оффлайн-доставки
      tags:
      - Messages
  /users/{userId}:
    get:
      description: Возвращает профиль пользователя вместе с его публичным identity-ключом
        (base64).
      parameters:
      - description: ID пользователя (UUID)
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.UserDTO'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить пользователя по ID
      tags:
      - Users
  /users/search:
    get:
      description: Возвращает пользователя по точному username. Поле found=false,
        если не найден.
      parameters:
      - description: username для поиска
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_user.SearchUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Поиск пользователя по username
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Формат: "Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"